deprecate advanced settings
- Move `export keystorе` functionality to Security and Privacy section.
- Remove advanced settings from menu

fixes #290
