+++ src/popup/router/pages/Settings.vue
@@ -11,11 +11,6 @@
                 <i class="arrowright"></i>
                 <p class="tabinfo">{{ $t('pages.settings.tabGeneralsmall') }}</p>
             </button>
-            <button class="advancedsett" @click="AdvancedSettings">
-                <span class="settings-li">{{ $t('pages.settings.tabAdvanced') }}</span>
-                <i class="arrowright"></i>
-                <p class="tabinfo">{{ $t('pages.settings.tabAdvancedsmall') }}</p>
-            </button>
             <button class="securitysett" @click="SecuritySettings">
                 <span class="settings-li">{{ $t('pages.settings.tabSecurity') }}</span>
                 <i class="arrowright"></i>
@@ -49,9 +44,6 @@ export default {
     methods: {
         GeneralSettings() {
             this.$router.push('/generalSettings')
-        },
-        AdvancedSettings() {
-            this.$router.push('/advancedSettings')
         },
         SecuritySettings() {
             this.$router.push('/securitySettings')
